improve map wrapping for private maps
