+++ app/maps/[slug]/layout.tsx
@@ -1,6 +1,7 @@
 import { Suspense } from 'react'
 import Loading from './loading'
 import { Metadata } from 'next'
+import Menu from '@/components/Menu'
 
 export const metadata: Metadata  = {
   openGraph: {
@@ -11,6 +12,7 @@ export const metadata: Metadata  = {
 export default function MapPageLayout({ children }: {children: React.ReactNode}) {
  return (
         <Suspense fallback={<Loading />}>
+          <Menu selectedPage={"maps"}></Menu>
             {children}
         </Suspense>
   )
